requirements.txt
torch~=2.4.1
torchvision~=0.19.1
streamlit~=1.39.0
pillow~=10.4.0
tqdm~=4.66.5